// src/app/shared/xapi/verb.enum.ts
export enum Verb {
    PersonCreated = 'http://eventuality.poc/xapi/verb/person/created',
    PersonCreationFailed = 'http://eventuality.poc/xapi/verb/person/creation-failed',
    PersonCreationRequested = 'http://eventuality.poc/xapi/verb/person/creation-requested',
    PersonRequested = 'http://eventuality.poc/xapi/verb/person/requested',
    PersonRetrievalFailed = 'http://eventuality.poc/xapi/verb/person/retrieval-failed',
    PersonRetrieved = 'http://eventuality.poc/xapi/verb/person/retrieved',
    PersonUpdated = 'http://eventuality.poc/xapi/verb/person/updated',
    PersonUpdateFailed = 'http://eventuality.poc/xapi/verb/person/updated-failed',
    PersonUpdateRequested = 'http://eventuality.poc/xapi/verb/person/update-requested',
    SystemExceptionOcurred = 'http://eventuality.poc/xapi/verb/system/exception-ocurred',
    SystemTest = 'http://eventualityPOC.com/xapi/verb/system/test'
}